--- h/api/groups/__init__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from h.api.groups.auth import group_principals
-from h.api.groups.auth import set_permissions
-from h.api.groups.logic import set_group_if_reply
-from h.api.groups.logic import insert_group_if_none
-
-__all__ = (
-    'group_principals',
-    'set_permissions',
-    'set_group_if_reply',
-    'insert_group_if_none'
-)
