+++ h/api/transform.py
@@ -1,37 +1,82 @@
 # -*- coding: utf-8 -*-
 
-"""Functions for transforming data for or from the search index."""
-
 from h._compat import string_types
 from h.api import nipsa
 from h.api import uri
-from h.api import groups
 
 
-def prepare(annotation):
+def set_group_if_reply(annotation, fetcher):
     """
-    Prepare the given annotation for indexing.
+    If the annotation is a reply set its group to that of its parent.
 
-    Scan the passed annotation for any target URIs or document metadata URIs
-    and add normalized versions of these to the document.
+    If the annotation is a reply to another annotation (or a reply to a reply
+    and so on) then it always belongs to the same group as the original
+    annotation. If the client sent any 'group' field in the annotation we will
+    just overwrite it!
     """
-    groups.set_group_if_reply(annotation)
-    groups.insert_group_if_none(annotation)
-    groups.set_permissions(annotation)
+    def is_reply(annotation):
+        """Return True if this annotation is a reply."""
+        if annotation.get('references'):
+            return True
+        else:
+            return False
 
-    # FIXME: Remove this in a month or so, when all our clients have been
-    # updated. -N 2015-09-25
-    _transform_old_style_comments(annotation)
+    if not is_reply(annotation):
+        return
 
-    # FIXME: When this becomes simply part of a search indexing operation, this
-    # should probably not mutate its argument.
-    _normalize_annotation_target_uris(annotation)
+    # Get the top-level annotation that this annotation is a reply
+    # (or a reply-to-a-reply etc) to.
+    top_level_annotation_id = annotation['references'][0]
+    top_level_annotation = fetcher(top_level_annotation_id)
 
-    if 'user' in annotation and nipsa.has_nipsa(annotation['user']):
-        annotation['nipsa'] = True
+    # If we can't find the top-level annotation, there's nothing we can do, and
+    # we should bail.
+    if top_level_annotation is None:
+        return
+
+    if 'group' in top_level_annotation:
+        annotation['group'] = top_level_annotation['group']
+    else:
+        if 'group' in annotation:
+            del annotation['group']
 
 
-def _normalize_annotation_target_uris(annotation):
+def insert_group_if_none(annotation):
+    if not annotation.get('group'):
+        annotation['group'] = '__world__'
+
+
+def set_group_permissions(annotation):
+    """Set the given annotation's permissions according to its group."""
+    # If this annotation doesn't have a permissions field, we don't know how to
+    # handle it and should bail.
+    permissions = annotation.get('permissions')
+    if permissions is None:
+        return
+
+    # For private annotations (visible only to the user who created them) the
+    # client sends just the user's ID in the read permissions.
+    is_private = (permissions.get('read') == [annotation['user']])
+
+    if is_private:
+        # The groups feature doesn't change the permissions for private
+        # annotations at all.
+        return
+
+    group = annotation.get('group')
+    if group == '__world__':
+        # The groups feature doesn't change the permissions for annotations
+        # that don't belong to a group.
+        return
+
+    group_principal = 'group:' + group
+
+    # If the annotation belongs to a group, we make it so that only users who
+    # are members of that group can read the annotation.
+    annotation['permissions']['read'] = [group_principal]
+
+
+def normalize_annotation_target_uris(annotation):
     if 'target' not in annotation:
         return
     if not isinstance(annotation['target'], list):
@@ -46,7 +91,7 @@ def _normalize_annotation_target_uris(annotation):
         target['scope'] = [uri.normalize(target['source'])]
 
 
-def _transform_old_style_comments(annotation):
+def fix_old_style_comments(annotation):
     """
     Transform an old-style, targetless "comment" into one with a target.
 
@@ -64,3 +109,8 @@ def _transform_old_style_comments(annotation):
 
     if has_uri and has_no_targets and has_no_references:
         annotation['target'] = [{'source': annotation.get('uri')}]
+
+
+def add_nipsa(annotation):
+    if 'user' in annotation and nipsa.has_nipsa(annotation['user']):
+        annotation['nipsa'] = True
